backend/requirements.txt
asgiref==3.7.2
astroid==3.0.1
black==23.11.0
click==8.1.7
coverage==7.3.2
dill==0.3.7
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
pylint==3.0.2
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.3
pusher==3.3.2